"more restructuring of MB files; updates #123"
