+++ prototype2/multiblade/generators/profiling.cpp
@@ -16,7 +16,7 @@
 #include <chrono>
 
 #include "generators/TextFile.h"
-#include "common/EventBuilder.h"
+#include "clustering/EventBuilder.h"
 
 // GCOVR_EXCL_START
